+++ src/service/architecture/application/index.ts
@@ -1,2 +1,3 @@
 export * from "./entry-points";
 export * from "./application";
+export * from "../../../dependency-injection";